requirements.txt
pyTelegramBotAPI==4.7.1
python-telegram-bot-calendar~=1.0.5
loguru~=0.6.0
peewee~=3.15.2
requests~=2.28.0
telebot~=0.0.4
python-dotenv~=0.21.0